requirements.txt
google-api-python-client==2.140.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
click==8.1.7
rich==13.7.1
python-dotenv==1.0.1
ollama==0.3.3
httpx==0.27.2